requirements.txt
pandas
numpy
langchain-community
pypdf
python-dotenv
langchain_openai
langchain_huggingface
langchain_groq
streamlit
faiss-cpu